Write `get_meshmanager` and use it in modules that need meshes
This allows caching of meshes to a folder on the user's computer,
preventing the need to re-download the same mesh many times. I should
probably add a warning somewhere if the user has downloaded a ton of
meshes, since it could take up a lot of disk space and the
saving-to-disk behavior is not super well documented currently so might
be an unwelcome surprise for some users who are low on disk space.