Add flexiblity to `annotations` argument
I've tried to save users the pain of building a complicated list of dicts
of strings and DataFrames, when they want to just do something simple
like show a set of points. Now users can provide numpy arrays of points
and the code takes care of building the list of dicts of strings and
DataFrames that it needs. I also implemented `annotation_units` so that
users can directly pass things like skeleton nodes which are in nm.