Update volume outlines layer to not try to load images
Oops, the volumes outline layer was configured in a way that
neuroglancer would try to pull image data from it, resulting in a ton of
requests being sent and failing to return anything. This jammed up the
network and caused image data to load slower, especially noticeable when
very zoomed out and trying to load low resolution image data of the
whole xy area of the VNC. After this commit, these unneeded requests for
image data from the volumes layer are gone, and EM image data loads
faster. Thanks to Nico for this solution – we had things this way for
FANC for quite a while but then I screwed it up when changing to pull
data from a Lee lab bucket instead of from a Zetta bucket, so this
commit reverts the configuration of the volume outlines layer to be more
like what Nico had originally made it as.