`lookup.py`: Standardized function names, variable names, docstrings
This is a major change, to be included in the version bump from 2.x.x to
3.0.0, because some function names changed, so users will need to update
function calls in their code. The function name changes are small (just
dropping an `s` from a few words that were previously plural – somas,
segids, pts), but still this is a non-backward-compatible change.

Other locations in this package that call these functions have been
changed.