Refactor `neuron_information_bot` to be able to post to multiple tables
The bot will check if the user's annotation is a valid annotation for
one of a short list of CAVE tables, and will upload to the first table
it finds for which the annotation is valid/accepted. This is a work in
progress and there may be some bugs left at this point, but the basic
functionality of querying and uploading works.