Fix a weird pd.concat bug. Improve an except statement.
pd.concat fails for weird reasons sometimes when the dataframes to be
concatenated have .attrs that are set in certain ways. It was failing
for me when I was trying to pull annotations from two different columns
of the same table. Unsetting the attrs fixed it. I don't think anyone
uses these attrs so I don't consider this a backward-incompatible
change, even though technically it could be if anyone used those.